# WIP:optimize CurrentMS with nanotime