name         old time/op  new time/op  delta
CurrentMS-4  81.1ns ± 2%  35.6ns ± 7%  -56.16%  (p=0.008 n=5+5)